Extract shared resolvePublicFile helper
Both server.ts and application.tsx independently defined
publicDirectoryUrls and file resolution logic. Extract into
a single resolve-public-file module to eliminate the duplication
and risk of drift.
